moved files to dir, made module, other changes
