+++ data/squares.py
